Fixed negative length issue decoding long LMBCS text list entries

+++ domino-jna/src/main/java/com/mindoo/domino/jna/internal/ItemDecoder.java
@@ -74,7 +74,7 @@ public class ItemDecoder {
 					listValues.add(lmbcsString);
 				}
 				else {
-					String currListEntry = NotesStringUtils.fromLMBCS(pointerToTextInMem, (short) retTextLengthAsInt);
+					String currListEntry = NotesStringUtils.fromLMBCS(pointerToTextInMem, retTextLengthAsInt);
 					listValues.add(currListEntry);
 				}
 			}
